functions/consolidate_labels/requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
functions-framework
